[Feat] Update based on comments
