+++ Week 01/Lecture 02/task7/src/main/java/utils/DateUtils.java
@@ -2,9 +2,10 @@ package utils;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
+import java.time.format.DateTimeParseException;
 
 public class DateUtils {
-    private static final DateTimeFormatter formatter = DateTimeFormatter.ofPattern("d/M/yyyy");
+    private static final DateTimeFormatter DATE_FORMATER = DateTimeFormatter.ofPattern("d/M/yyyy");
 
     /**
      * Parses a date string in the format "d/M/yyyy" and returns a LocalDate object.
@@ -13,8 +14,13 @@ public class DateUtils {
      * @return the parsed LocalDate object
      * @throws IllegalArgumentException if the date string cannot be parsed
      */
-    public static LocalDate parseDate(String dateString) {
-        return LocalDate.parse(dateString, formatter);
+    public static LocalDate parseDate(String dateStr) {
+        try {
+            return LocalDate.parse(dateStr, DATE_FORMATER);
+        } catch (DateTimeParseException e) {
+            System.out.println("Error parsing date: " + dateStr);
+            throw e;
+        }
     }
 
     /**
@@ -24,6 +30,6 @@ public class DateUtils {
      * @return the formatted string in the specified format
      */
     public static String formatDate(LocalDate date) {
-        return date.format(formatter);
+        return date.format(DATE_FORMATER);
     }
 }
