+++ Week 03/Lecture 05/Assignment 02/lecture_5_hw/src/main/java/com/example/lecture_5_hw/utils/FileUtils.java
@@ -3,6 +3,7 @@ package com.example.lecture_5_hw.utils;
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
+import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -25,12 +26,27 @@ public class FileUtils {
             br.readLine(); // Skip header
             while ((line = br.readLine()) != null) {
                 String[] attributes = line.split(",");
-                Employee employee = Employee.fromCSV(attributes);
+                Employee employee = fromCSV(attributes);
                 employees.add(employee);
             }
         } catch (IOException e) {
             throw new IOException("Error reading employee (Manual) " + e);
         } 
         return employees;
+    }
+
+    /**
+     * Parses an array of attributes into an Employee object.
+     *
+     * @param attributes an array of strings representing the employee's id, name, date of birth, address, and department.
+     * @return an Employee object created from the provided attributes.
+     */
+    public static Employee fromCSV(String[] attributes) {
+        String id = attributes[0];
+        String name = attributes[1];
+        LocalDate dob = DateUtils.parseDate(attributes[2]);
+        String address = attributes[3];
+        String department = attributes[4];
+        return new Employee(id, name, dob, address, department);
     }
 }
